Refactor applications_api folder structure
